Moved all icons of toolbar to commonMain and accessing the resource in commonMain
