+++ jrteLib/build.gradle.kts
@@ -32,9 +32,8 @@ kotlin {
                 api(compose.runtime)
                 api(compose.foundation)
                 api(compose.material3)
-                api("androidx.core:core-ktx:1.9.0")
-                implementation("org.jetbrains.kotlin:kotlin-reflect:$kotlin_version")
                 api("com.arkivanov.essenty:parcelable:0.10.0")
+                implementation("org.jetbrains.compose.components:components-resources:1.4.0-alpha01-dev942")
             }
         }
 
@@ -44,11 +43,7 @@ kotlin {
             }
         }
 
-        val androidMain by getting {
-            dependencies {
-                api("androidx.core:core-ktx:1.9.0")
-            }
-        }
+        val androidMain by getting
         val androidTest by getting
         val iosX64Main by getting
         val iosArm64Main by getting
@@ -74,8 +69,9 @@ kotlin {
 android {
     namespace = "com.jrte"
     compileSdk = 33
+    sourceSets["main"].resources.srcDirs("src/commonMain/resources")
     defaultConfig {
-        minSdk = 23
+        minSdk = 24
         targetSdk = 33
     }
 }
