+++ jrteLib/src/commonMain/kotlin/com/jrte/ui/composable/EmptyContextMenuToolbar.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.geometry.Rect
 import androidx.compose.ui.platform.TextToolbar
 import androidx.compose.ui.platform.TextToolbarStatus
 
-object EmptyTextToolbar: TextToolbar {
+object EmptyContextMenuToolbar: TextToolbar {
     override val status: TextToolbarStatus = TextToolbarStatus.Hidden
 
     override fun hide() {  }
