+++ jrteLib/src/commonMain/kotlin/com/jrte/ui/composable/JRTEditor.kt
@@ -1,18 +1,25 @@
 package com.jrte.ui.composable
 
-import androidx.compose.runtime.Composable
+import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.WindowInfo
 import com.jrte.ui.textstyle.RichTextValue
 
 @Composable
 fun JRTEditor(
-    value: RichTextValue,
-    onValueChange: (RichTextValue) -> Unit,
     modifier: Modifier = Modifier,
     textFieldStyle: RichTextFieldStyle = defaultRichTextFieldStyle(),
     readOnly: Boolean = false,
 ) {
+    var value by remember {
+        mutableStateOf(
+            RichTextValue.fromString(
+                // Optional parameter; leave it blank if you want to use provided styles
+                // But if you want to customize the user experience you're free to do that
+                // by providing a custom StyleMapper
+                styleMapper = CustomStyleMapper()
+            )
+        )
+    }
     RichTextField(
         modifier = modifier,
         value = value.value,
@@ -21,7 +28,7 @@ fun JRTEditor(
         onValueChange = {
             val newValue = value.copy()
             if (newValue.updatedValueAndStyles(it)) {
-                onValueChange(newValue)
+                value = newValue
             }
         },
         textFieldStyle = textFieldStyle,
