+++ jrteLib/src/commonMain/resources/icon_bold.xml
