+++ jrteLib/src/commonMain/resources/icon_italic.xml
